+++ packages/core/src/presentation/webview/utils/formatters/index.ts
@@ -13,3 +13,4 @@ export * from './styleFlagsFormatter';
 export * from './wordFrequencyFormatter';
 export * from './categorySearchFormatter';
 export * from './analysisFormatter';
+export * from './streamingStatsFormatter';